+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.3.1] 2022-10-31
+### Added
+- `AudioCaps.load_class_labels_indices` to load AudioSet classes map externally.
+
 ### Changed
 - AudioCaps, Clotho and MACS order are now defined by their order in the corresponding captions CSV files when available.
 - Update documentation usage and main page.